requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.11.0
exceptiongroup==1.1.3
fonttools==4.42.1
h11==0.14.0
humanfriendly==10.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
outcome==1.2.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pycparser==2.21
pyparsing==3.1.1
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.5
verboselogs==1.7
webdriver-manager==4.0.1
wsproto==1.2.0
